💄 공통 스타일 모듈 통합
