+++ app/frontend/src/pages/Profile/MyMogaco/index.css.ts
@@ -1,28 +1,12 @@
 import { style } from '@vanilla-extract/css';
 
 import { vars } from '@/styles';
-import { sansBold16, sansRegular16 } from '@/styles/font.css';
+import { sansRegular16 } from '@/styles/font.css';
 
 import * as parentStyle from '../index.css';
 
-export const { section, list } = parentStyle;
-
-export const loading = style({
-  width: '100%',
-  display: 'flex',
-  justifyContent: 'center',
-});
-
-export const mogacoListButton = style([
-  sansBold16,
-  {
-    display: 'flex',
-    gap: 0,
-    alignItems: 'center',
-    alignSelf: 'end',
-    color: vars.color.morakGreen,
-  },
-]);
+export const { section, list, loading, navLinkButton, rotateArrow } =
+  parentStyle;
 
 export const notParticipated = style([
   sansRegular16,
@@ -33,7 +17,3 @@ export const notParticipated = style([
     color: vars.color.grayscale200,
   },
 ]);
-
-export const rotateArrow = style({
-  transform: 'rotate(180deg)',
-});
