+++ app/frontend/src/pages/Profile/index.css.ts
@@ -1,5 +1,8 @@
 import { style } from '@vanilla-extract/css';
 
+import { vars } from '@/styles';
+import { sansBold16 } from '@/styles/font.css';
+
 export const container = style({
   display: 'flex',
   flexDirection: 'column',
@@ -14,6 +17,27 @@ export const list = style({
   gap: '1.6rem',
   maxHeight: '48rem',
   overflowY: 'auto',
+});
+
+export const loading = style({
+  width: '100%',
+  display: 'flex',
+  justifyContent: 'center',
+});
+
+export const navLinkButton = style([
+  sansBold16,
+  {
+    display: 'flex',
+    gap: 0,
+    alignItems: 'center',
+    alignSelf: 'end',
+    color: vars.color.morakGreen,
+  },
+]);
+
+export const rotateArrow = style({
+  transform: 'rotate(180deg)',
 });
 
 export const section = style({
